"ajout delete sur le manual et route necessaires + retouches bootstrap"

+++ src/Controller/ProductController.php
@@ -51,15 +51,15 @@ class ProductController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            
+
             $product = $form->getData();
-            
+
             $purchaseDate = $product->getPurchaseDate();
             $warrantyDuration = $product->getWarrantyDuration();
             $expirationDate = clone $purchaseDate;
             $expirationDate = $expirationDate->add(new DateInterval('P' . $warrantyDuration . 'Y'));
             $product->setExpirationDate($expirationDate);
-            
+
             $productRepository->save($product, true);
 
             return $this->redirectToRoute('app_product_index', [], Response::HTTP_SEE_OTHER);
@@ -79,11 +79,11 @@ class ProductController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            
+
             $productManual = $form->getData();
-            
-            
-            $productManual->setProduct($product);            
+
+
+            $productManual->setProduct($product);
             $productDocRepository->save($productManual, true);
 
             return $this->redirectToRoute('app_product_index', [], Response::HTTP_SEE_OTHER);
@@ -92,7 +92,7 @@ class ProductController extends AbstractController
         return $this->renderForm('product/addManual.html.twig', [
             'product' => $product,
             'form' => $form,
-            
+
         ]);
     }
 
@@ -102,7 +102,7 @@ class ProductController extends AbstractController
 
         $form = $this->createForm(CollecType::class);
         $form->handleRequest($request);
-        
+
 
         if ($form->isSubmitted() && $form->isValid()) {
             foreach ($form->get('docs')->getData() as $doc) {
@@ -133,15 +133,15 @@ class ProductController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            
+
             $product = $form->getData();
-            
+
             $purchaseDate = $product->getPurchaseDate();
             $warrantyDuration = $product->getWarrantyDuration();
             $expirationDate = clone $purchaseDate;
             $expirationDate = $expirationDate->add(new DateInterval('P' . $warrantyDuration . 'Y'));
             $product->setExpirationDate($expirationDate);
-            
+
             $productRepository->save($product, true);
 
             return $this->redirectToRoute('app_product_index', [], Response::HTTP_SEE_OTHER);
@@ -156,33 +156,53 @@ class ProductController extends AbstractController
     #[Route('/{id}', name: 'app_product_delete', methods: ['POST'])]
     public function delete(Request $request, Product $product, ProductRepository $productRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$product->getId(), $request->request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete' . $product->getId(), $request->request->get('_token'))) {
             $productRepository->remove($product, true);
         }
 
         return $this->redirectToRoute('app_product_index', [], Response::HTTP_SEE_OTHER);
     }
 
-    #[Route('/delete-doc{id}', name: 'app_product_delete', methods: ['POST', 'DELETE'])]
+    #[Route('/delete-doc{id}', name: 'app_productDoc_delete', methods: ['POST', 'DELETE'])]
     public function deleteDoc(int $id, ProductDocRepository $productDocRepository)
     {
-    // Recherche du productDoc par ID
-    $productDoc = $productDocRepository->find($id);
+        // Recherche du productDoc par ID
+        $productDoc = $productDocRepository->find($id);
 
-    // Suppression du productDoc
-    $productDocRepository->remove($productDoc, true);
+        // Suppression du productDoc
+        $productDocRepository->remove($productDoc, true);
 
-    // Vérification que le produit existe
-    if (!$productDoc) {
-        return new JsonResponse(['message' => 'Doc non trouvé'], 404);
+        // Vérification que le produit existe
+        if (!$productDoc) {
+            return new JsonResponse(['message' => 'Doc non trouvé'], 404);
+        }
+
+        // Suppression du produit
+        // $product->delete();
+
+        // Retour d'une réponse de succès
+        return new JsonResponse(['message' => 'Doc supprimé avec succès']);
     }
 
-    // Suppression du produit
-    // $product->delete();
+    #[Route('/delete-manual/{id}', name: 'app_product_deleteManual', methods: ['POST', 'DELETE'])]
+    public function deleteManual(int $id, ProductDocRepository $productDocRepository, Request $request)
+    {
+        // Recherche du productDoc par ID
+        $productDoc = $productDocRepository->find($id);
 
-    // Retour d'une réponse de succès
-    return new JsonResponse(['message' => 'Doc supprimé avec succès']);
+        // Vérification que le productDoc existe et a un champ manual
+        if (!$productDoc || !$productDoc->getManual()) {
+            return new JsonResponse(['message' => 'Manual non trouvé'], 404);
+        }
+
+        // Suppression du manual
+        $productDocRepository->remove($productDoc, true);
+
+        // Retour d'une réponse de succès
+        $referer = $request->headers->get('referer');
+        return $this->redirect($referer);
     }
+
 
 
 
@@ -205,6 +225,4 @@ class ProductController extends AbstractController
             'products' => $products,
         ]);
     } */
-
-    
 }
